feat(Profile Window): OZF-573 Profile Window
MESSAGE FOR PREVIOUS MERGE:

A new UI component has been added - a modal window for viewing basic
profile information.  The window displays the user's username, display
name, email, and list of owned listings (as links).  This window can be
accessed via the global nav menu, the listing details tab, and
changelog entries.  When the window is closed, the previous modal
opens, if there is one.

Breaking changes:
Build changes: The build now required npm 2, as the shrinkwrap
functionality in npm 1 for some reason could not handle the new
ozp-react-commons repository