feat(folders): Support moving bookmarks into folders OZBE7
OZBE7 = ozone-development/ozp-backend#7

There are a couple of thing with the change in LibraryStore worth
noting. Prior to this new backend change the api call /library was used
for both gets and puts. When used to update that endpoint returned data,
data that was used in the updateLibrary function. With the new backend's
/update_all endpoint, that data is no longer returned, so I just forward
libraryEntries onto updateLibrary.

The second thing to note has to do with timing (race condition?), and
why I had to add a LibraryApi.get() to onUpdateLibrary(). HUD, to keep
it synced with Center, calls fetchLibrary (makes an API call) whenever
there is a focus event. When updating the folder name in the folder
modal, both onFetchLibrary() and onUpdateLibrary() are called on
focusing. With the new backend, but not the old one, the updateLibrary()
call returns before the onFetchLibrary() returns. Put another way, the
PUT completes before the GET completes. React properly renders the data
after the PUT and shortly thereafter re-renders now-stale data from the
GET. You can avoid this using a short setTimeout period to delay calling
updateLibrary, but determining how long the delay could be difficult and
or network dependent. Instead, I've inserted a call to LibraryApi.get()
which should take about the same amount of time as the call in
onFetchLibrary (the Center-syncing one) but start later (after the PUT),
hopefully forestalling the race.